rename modules & window RELOAD button
