+++ app/src/renderer/Player/WebPlayer.vue
